Added new behaviours for 2 powerups
The "rocket" powerup allows to jump high for 15 seconds now, and the
"star/fairy" powerup allows free flying for 5 seconds now. Change outfit
when collected is still missing. Thinking about a "powerup inventory" so
player can use powerups when he/she decides instead of the moment when
powerups are collected.