New things 03/12/2015
I have changed my mind, now the power ups are collected and the used
pressing different keys. I started to work in the enemies as well, since
I need enemies to develope the dinosaur power up. I have to think how to
change the behaviour of the enemies when  Rosie is in dinosaur mode.