Everything is working!
At least everything implemented so far. The dinosaur power up has
changed a little bit. Rosie-dinosaur breathes fire now. This way killing
only certain instances of bad doggy is easier, it just kills the objects
that are overlapping the fire. Pressing G to activate the dinomorph
Rosie grants 1 second of fire coming out her mouth. With this change,
all the core elements of the game are working. I'll think during this
days whether I create other types of enemies. The other thing I think
about is a progress bar to show the duration of the power ups. I don't
know if the players will find the timing very difficult and obscure
without the bar or quite the opposite, a bar would suppress the thrill
in the game. I'll ask my testers about that