Fire, Levels and shooting dogs
First thing done today was making sure that the fire Rosie breaths cames
out from the correct side of her head.
After that, I created a whole new level and a layout to select levels.
The last thing was creating a third level, but I decided to create a new
enemy for this level, a shoothing doggie. Y created the bullets and the
dog, but it took a long time there is still not a third level.