Added 3rd level
The new level has the shooting dog and a highest level of difficulty.
Also I have discovered that once I forget how the level is supposed to
be overcomed it becomes more difficult, so the level puzzle is not as
obvious as a thought. Also the game is having some performance trouble
and I don't know why this could be happening.